add unit of work pattern

+++ ChatApp/Hubs/ChatHub.cs
@@ -1,9 +1,7 @@
 ﻿
-using AutoMapper;
-using ChatApp.Data;
-using ChatApp.Data.Repository.Users;
+using ChatApp.Data.Modals;
+using ChatApp.Data.UnitOfWork;
 using ChatApp.Models;
-using ChatApp.Models.DTOs;
 using ChatApp.Services;
 using ChatApp.Services.ChatServices;
 using Microsoft.AspNetCore.SignalR;
@@ -13,29 +11,32 @@ namespace ChatApp.Hubs
 {
     public class ChatHub : Hub
     {
-        private readonly IUserRepository _userRepository;
         private readonly ChatAppDBContext _dbContext;
+        private readonly IUnitOfWork _unitOfWork;
         private readonly ICacheService _cacheService;
         private readonly IHubService _hubService;
 
-        public ChatHub(IUserRepository userRepository, ChatAppDBContext dbContext, ICacheService cacheService,  IHubService hubService) : base()
+        public ChatHub(IUnitOfWork unitOfWork, ICacheService cacheService,  IHubService hubService, ChatAppDBContext chatAppDBContext) : base()
         {
-            _userRepository = userRepository;
-            _dbContext = dbContext;
+            _unitOfWork = unitOfWork;
             _cacheService = cacheService;
             _hubService = hubService;
+            _dbContext = chatAppDBContext;
         }
         public override async Task OnDisconnectedAsync(Exception? exception)
         {
             var usersConn = await _cacheService.GetDataByEndpoint<UserConnection>("list-users-online");
-            var current = usersConn.Where(user => user.ConnectionId == Context.ConnectionId).FirstOrDefault();
-            if (current != null)
+            if (usersConn != null)
             {
-                await _userRepository.UpdateStatusOnline(current.UserId, false);
-                // remove cache data
-                await _cacheService.RemoveDataByKey($"list-users-online:{current.UserId}");
-                // return list users online
-                await _hubService.NotifyFriendsOnline(current.UserId);
+                var current = usersConn.Where(user => user.ConnectionId == Context.ConnectionId).FirstOrDefault();
+                if (current != null)
+                {
+                    await _unitOfWork.UserRepository.UpdateStatusOnline(current.UserId, false);
+                    // remove cache data
+                    await _cacheService.RemoveDataByKey($"list-users-online:{current.UserId}");
+                    // return list users online
+                    await _hubService.NotifyFriendsOnline(current.UserId);
+                }
             }
         }
         public async Task AddUserConnection(UserConnection user)
@@ -48,7 +49,7 @@ namespace ChatApp.Hubs
                 Avatar = user.Avatar ?? "",
             };
             // update database
-            await _userRepository.UpdateStatusOnline(user.UserId, true);
+            await _unitOfWork.UserRepository.UpdateStatusOnline(user.UserId, true);
             // add user in cache
             await _cacheService.SetData($"list-users-online:{userConn.UserId}", userConn, TimeSpan.FromHours(24));
             // join group chat
